+++ src/components/settings-ui/index.ts
@@ -1,2 +1,3 @@
 export { AccountSection } from "./AccountSection";
 export { DeleteAccountDialog } from "./DeleteAccountDialog";
+export { CoverBackfill } from "./CoverBackfill";